version2/hackathon-work/requirements.txt
flask
solana
solders
python-dotenv
requests
elevenlabs
base58